Time: 27 ms (48.07%), Space: 96 MB (17.99%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -14,7 +14,7 @@ public:
         sort(merged.begin(), merged.end());
         int total = merged.size();
         if(total % 2 == 1){
-            return static_cast<double>(merged[total/2]);
+            return (merged[total/2]);
         }
         else{
             int middle1 = merged[total/2 -1];
